refactor by spliitting on files; add task id, status, list
