+++ python/src/clients/serviceinfo.rs
@@ -1,0 +1,32 @@
+use pyo3::prelude::*;
+use ga4gh_sdk::clients::serviceinfo::ServiceInfo;
+use crate::PyConfiguration;
+
+#[pyclass(name = "ServiceInfo", module = "GA4GH")]
+pub struct PyServiceInfo {
+    inner: ServiceInfo,
+}
+
+#[pymethods]
+impl PyServiceInfo {
+    #[new]
+    pub fn new(py_config: &PyConfiguration) -> PyResult<Self> {
+        let service_info = ServiceInfo::new(&py_config.inner).unwrap();
+        Ok(PyServiceInfo {
+            inner: service_info,
+        })
+    }
+
+    pub fn get_service_info(&self) -> PyResult<String> {
+        let rt = tokio::runtime::Runtime::new().unwrap();
+        let result = rt.block_on(self.inner.get());
+
+        match result {
+            Ok(service_info) => Ok(format!("Service: {:?}", service_info)),
+            Err(e) => Err(pyo3::exceptions::PyRuntimeError::new_err(format!(
+                "Failed to get service info: {}",
+                e
+            ))),
+        }
+    }
+}
